backend/requirements.txt
fastapi>=0.68.0
uvicorn[standard]>=0.15.0
psycopg2-binary>=2.9.0
python-dotenv>=0.19.0

